% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_menu_info.R
\name{get_menu_info}
\alias{get_menu_info}
\title{Get the nutritional information of menu items.}
\usage{
get_menu_info(key = NULL, item_id = NULL)
}
\arguments{
\item{key}{API key}

\item{item_id}{The menu item ID.}
}
\value{
The nutritional information of menu items.
}
\description{
This functions gets the nutritional information of one menu item using menu item ID.
}
\examples{
get_menu_info(key = Sys.getenv("SPOON_KEY"), item_id = 278131)
}
