% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe.R
\name{get_recipe}
\alias{get_recipe}
\title{Get the recipes.}
\usage{
get_recipe(
  key = NULL,
  query = NULL,
  cuisine = NULL,
  diet = NULL,
  exclude = NULL,
  intolerance = NULL,
  number = 10,
  instruction = NULL,
  maxReadyTime = NULL,
  minCalories = NULL,
  maxCalories = NULL,
  minCarbs = NULL,
  maxCarbs = NULL,
  minProtein = NULL,
  maxProtein = NULL,
  minFat = NULL,
  maxFat = NULL,
  minAlcohol = NULL,
  maxAlcohol = NULL,
  minCaffeine = NULL,
  maxCaffeine = NULL,
  minSaturatedFat = NULL,
  maxSaturatedFat = NULL
)
}
\arguments{
\item{key}{API key}

\item{query}{The (natural language) recipe search query.}

\item{cuisine}{The cuisine(s) of the recipes. One or more, comma separated (will be interpreted as 'OR')}

\item{diet}{The diet for which the recipes must be suitable.}

\item{exclude}{A comma-separated list of ingredients or ingredient types that the recipes must not contain.}

\item{intolerance}{A comma-separated list of intolerances. All recipes returned must not contain ingredients that are not suitable for people with the intolerances entered.}

\item{number}{The number of expected results (between 1 and 10).}

\item{instruction}{Whether the recipes must have instructions.}

\item{maxReadyTime}{The maximum time in minutes it should take to prepare and cook the recipe.}

\item{minCalories}{The minimum amount of calories the recipes must have.}

\item{maxCalories}{The maximum amount of calories the recipes can have.}

\item{minCarbs}{The minimum amount of carbohydrates in grams the recipes must have.}

\item{maxCarbs}{The maximum amount of carbohydrates in grams the recipes can have.}

\item{minProtein}{The minimum amount of protein in grams the recipes must have.}

\item{maxProtein}{The maximum amount of protein in grams the recipes can have.}

\item{minFat}{The minimum amount of fat in grams the mrecipes must have.}

\item{maxFat}{The maximum amount of fat in grams the recipes can have.}

\item{minAlcohol}{The minimum amount of alcohol in grams the recipe must have.}

\item{maxAlcohol}{The maximum amount of alcohol in grams the recipe can have.}

\item{minCaffeine}{The minimum amount of caffeine in milligrams the recipe must have.}

\item{maxCaffeine}{The maximum amount of caffeine in milligrams the recipe can have.}

\item{minSaturatedFat}{The minimum amount of saturated fat in grams the recipe must have.}

\item{maxSaturatedFat}{The maximum amount of saturated fat in grams the recipe can have.}
}
\value{
The recipes requested by users.
}
\description{
This functions gets the recipes with specified requirements.
}
\examples{
get_recipe(key = Sys.getenv("SPOON_KEY"), query = "pork" ,number = 10)
}
