% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe_equipment.R
\name{get_recipe_equipment}
\alias{get_recipe_equipment}
\title{Get the equipments required in recipes.}
\usage{
get_recipe_equipment(key = NULL, recipe_id = NULL)
}
\arguments{
\item{key}{API key}

\item{recipe_id}{The recipe ID.}
}
\value{
The equipments for the recipe entered.
}
\description{
This functions gets the equipments needed for one recipe.
}
\examples{
get_recipe_equipment(key = Sys.getenv("SPOON_KEY"), recipe_id = "753644")
}
