% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe_information.R
\name{get_recipe_information}
\alias{get_recipe_information}
\title{Get detailed information of multiple recipes.}
\usage{
get_recipe_information(key = NULL, recipe_ids = NULL)
}
\arguments{
\item{key}{API key}

\item{recipe_ids}{The recipe IDs.}
}
\value{
The detailed information of the recipes entered.
}
\description{
This functions gets detailed information of multiple recipes entered.
}
\examples{
get_recipe_information(key = Sys.getenv("SPOON_KEY"), recipe_ids = "509163,825127,949313")
}
