% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe_ingredient.R
\name{get_recipe_ingredient}
\alias{get_recipe_ingredient}
\title{Get the ingredients required in recipes.}
\usage{
get_recipe_ingredient(key = NULL, recipe_id = NULL)
}
\arguments{
\item{key}{API key}

\item{recipe_id}{The recipe ID.}
}
\value{
The ingredients for the recipe entered.
}
\description{
This functions gets ingredients needed for one recipe.
}
\examples{
get_recipe_ingredient(key = Sys.getenv("SPOON_KEY"), recipe_id = "753644")
}
