% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe_nutrition.R
\name{get_recipe_nutrition}
\alias{get_recipe_nutrition}
\title{Get the nutritional information of the recipes.}
\usage{
get_recipe_nutrition(key = NULL, recipe_id = NULL)
}
\arguments{
\item{key}{API key}

\item{recipe_id}{The recipe ID.}
}
\value{
The nutritional information of the recipe entered.
}
\description{
This functions gets nutritional information of one recipe.
}
\examples{
get_recipe_nutrition(key = Sys.getenv("SPOON_KEY"), recipe_id = "753644")
}
